Constructors for the quiz

Construct a quiz or quiz question

Usage

construct_quiz(..., options = set_quiz_options())

construct_question(
  prompt,
  answerUserPrettifier,
  answerCorrectPretty,
  grader,
  ns
)

construct_messages(message_correct, message_wrong, message_skipped)

Arguments

...

objects of class 'quizQuestions'. See construct_question()

options

a list of options generated from set_quiz_options()

prompt

an htmltools::div that represents a quiz question

answerUserPrettifier

a function that takes the user answer and prints it neatly. This is wrapped with purrr::possibly() to catch any errors.

answerCorrectPretty

a character that prints the correct answer neatly

grader

a function that takes the user answer and determines if it is correct. Must take one argument and return TRUE or FALSE. This is wrapped with purrr::possibly() and base::isTRUE() to catch any errors.

ns

namespace generated from shiny::NS()

message_correct

a string to be shown at the end of the quiz when the user gets all questions correct

message_wrong

a string to be shown at the end of the quiz when the user gets at least one question wrong

message_skipped

a string to be shown at the end of the quiz when the user skips the quiz or ends it early

Value

an object of class quiz

an object of class quizQuestion

an object of class quizMessages

Details

See dev/example-app.R for current example.

Functions

construct_quiz(): Construct a quiz object

construct_question(): Construct a question object

construct_messages(): Construct a messages object

See also

Author

Joseph Marlo