Functions for managing the quiz state machine

The quiz is implemented via a state machine framework. It has states for each question and a final state for once the quiz ends. Only one state can be active at a time. The question text and answers shown depend on which state is active.

Usage

sm_get_state(store, variable = NULL, state = NULL)

sm_set_state(store, variable, value, state = NULL)

sm_is_current_correct(store)

sm_check_is_each_correct(store)

sm_is_all_correct(store)

sm_quiz_in_sandbox_mode(store)

sm_quiz_is_complete(store)

sm_logic_end_on_first_wrong(store)

sm_ui_format_prompts(quiz)

sm_ui_format_prompt(prompt, i)

sm_ui_quiz_complete(store, ns, messages)

sm_ui_complete_report(store)

sm_ui_question(store, ns)

sm_show_progress(store)

sm_resample_questions_if_sandbox(quiz, n = 50)

sm_create_reactive_store(quiz)

Arguments

store

a list formatted like in the example

variable

one of c('current-question', 'current-correct-answer', 'next-state', 'current-response')

state

one of c('quiz-question-1', ..., 'quiz-question')

quiz

an object of class 'quiz'

n

Number of resamples to make

Value

depends on function

questions

reactiveValues

Details

These are get and set functions for retrieving state values and setting values. The states are originally created via a reactiveValues call within Shiny server (or list outside of Shiny; see example below).

Functions

sm_get_state(): Getter function for the state machine

sm_set_state(): Setter function for the state machine

sm_is_current_correct(): Check that current-response is correct

sm_check_is_each_correct(): Check that each recorded answer is correct and return a boolean vector

sm_is_all_correct(): Check that all recorded answers are correct

sm_quiz_in_sandbox_mode(): Check if the quiz in sandbox mode

sm_quiz_is_complete(): Check if the quiz is complete

sm_logic_end_on_first_wrong(): Check if the quiz should end early if user fails a question

sm_ui_format_prompts(): Add headers containing the question number to all the questions in a quiz

sm_ui_format_prompt(): Add a header denoting the question number

sm_ui_quiz_complete(): UI to show once the quiz is completed

sm_ui_complete_report(): UI to show the score and table of correct answers to display at the end

sm_ui_question(): UI to show for the current question

sm_show_progress(): Show the progress bar if stipulated

sm_resample_questions_if_sandbox(): Create quasi infinite quiz by resampling questions n times

sm_create_reactive_store(): Create the main store object that handles the state(s)

See also

Author

Joseph Marlo

Examples

if (FALSE) {
quiz <- create_quiz(
  create_question(
    'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Select nulla.',
    add_choice('auctor'),
    add_choice('nulla', correct = TRUE)
  ),
  create_question(
    'Mauris congue aliquet dui, ut dapibus lorem porttitor sed. Select 600.',
    add_choice('600', correct = TRUE),
    add_choice('800')
  )
)
store <- sm_create_reactive_store(quiz)
sm_get_state(store)
sm_get_state(store, 'next-state')
sm_get_state(store, 'current-question')
sm_check_is_each_correct(store)
sm_quiz_in_sandbox_mode(store)
}